==> Dolgos_PDOs-Dolgos_Tissue_Matching/3_broad_cell_type_annotation.Rmd <==
---
title: "3.Dolgos patient matching tissue & PDOs: broad cell type assignation"
author: "Romuald Parmentier"
date: "2024-10-18"
output: html_document
---

```{r Load libraries and create output folder, message = F}

# Libraries
library(Seurat)
library(SingleCellExperiment)
library(ggplot2)
library(dplyr)
library(SingleR) # Cell type annotation
library(celldex)

# Functions and palettes
source(paste0(github_dir, "/Dolgos_Custom_Functions.R"))
source(paste0(github_dir, "/Dolgos_Custom_Color_Palettes.R"))

# Create output path
out_path = create_exp_folder(
  github_dir = github_dir,
  samples_ID = "Dolgos_PDOs-Dolgos_Tissue_Matching",
  exp = "broad_cell_type_annotation"
)

```

# Prepare the data

```{r Load the files}

# Load rescaled atlas sce object after dimension reduction
file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "Dolgos_2024_Matched_Tissue_PDOs",
  prev_exp = "dimension_reduction", 
  pattern = "sce_comb_dim_reduced.rds")

sce_comb = readRDS(file = file_path) 

# Load BlueprintEncodeData ref for broad cell type assignation
file_path = "/scicore/home/wykopa75/GROUP/rparmentier/sc_RNAseq/Projects/data/BlueprintEncodeData/BlueprintEncodeData.rds"

Blueprint_ref = readRDS(file = file_path) 

```

```{r Restrict sce object to published datasets}

# Restrict all_datasets sce object to Huang/Song/McCray as only those cells will be assigned thanks to singleR

indices_tissue = grep(
  x = sce_comb$Organoid_Type,
  pattern = "parental_tumor")

sce_comb_tissue = sce_comb[,indices_tissue]

```

# Broad cell type assignation

```{r Using Blueprint ref to assign broad cell types}
# Compute the prediction
pred_type <- SingleR(
  test = sce_comb, 
  ref = Blueprint_ref, 
  labels = Blueprint_ref$label.main)

# Assign braod cell types to tissues
sce_comb$broad_type = pred_type$pruned.labels

```


```{r Rearrange broad types according names}

# Simplify broad type catagories
sce_comb$broad_type = ifelse(
  test = sce_comb$broad_type %in% c("B-cells","CD4+ T-cells","CD8+ T-cells","Eosinophils","HSC","Macrophages", "Monocytes","NK cells"),
  yes = "Immune cells",
  no =  sce_comb$broad_type)


```


# Export files

```{r Export sce_comb object}

saveRDS(
  object = sce_comb,
  file = paste0(out_path,time_stamp(),"sce_comb_broad_cell_type_annotated.rds"))

```
